requirements.txt
Flask
gunicorn
watson-developer-cloud
ibm-watson
requests
python-dotenv
flask-cors